atomic-examples/deep-research/pyproject.toml
[tool.poetry]
name = "deep-research"
version = "0.1.0"
description = ""
authors = ["KennyVaneetvelde <kenny@inosta.be>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
atomic-agents = {path = "../.."}
requests = "^2.32.3"
beautifulsoup4 = "^4.12.3"
markdownify = "^0.13.1"
readability = "^0.3.1"
readability-lxml = "^0.8.1"
lxml-html-clean = "^0.4.0"
lxml = "^5.3.0"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
